PR feedback, refactor classes into separate files, fix missing paren
